- fixed even more PMD warnings and refactored some code
git-svn-id: http://mapsforge.googlecode.com/svn/trunk@1273 a424f762-2dd8-11df-bfeb-7b4c664f286a

+++ mapsforge/src/org/mapsforge/android/maps/Deserializer.java
@@ -70,7 +70,7 @@ final class Deserializer {
 	}
 
 	/**
-	 * Converts two bytes of a byte array to a signed short.
+	 * Converts two bytes of a byte array to a signed int.
 	 * <p>
 	 * The byte order is big-endian.
 	 * 
@@ -78,10 +78,10 @@ final class Deserializer {
 	 *            the byte array.
 	 * @param offset
 	 *            the offset in the array.
-	 * @return the short value.
+	 * @return the int value.
 	 */
-	static short getShort(byte[] buffer, int offset) {
-		return (short) (buffer[offset] << 8 | (buffer[offset + 1] & 0xff));
+	static int getShort(byte[] buffer, int offset) {
+		return buffer[offset] << 8 | (buffer[offset + 1] & 0xff);
 	}
 
 	/**
